Treat softbreak as a space
In order to wrap long paragraphs of markdown prose, it's reasonable to
wrap content on many lines and represent softbreak as a space in the
content. Otherwise it may combine text without spaces, creating a
smattering of joinedwords in the output. Yes, joinedwords was on purpose
to illustrate what the problem output could look like.